requirements.txt
webdriver-manager~=4.0.2
selenium~=4.33.0
pyyaml~=6.0.2
nicegui~=2.20.0